feat: device preference under profile
