--- app/models/github-integration.js
@@ -1,7 +1,0 @@
-import DS from 'ember-data';
-
-export default DS.Model.extend({
-
-  githubRepo: DS.attr('string')
-
-});
