--- app/models/jira-integration.js
@@ -1,7 +1,0 @@
-import DS from 'ember-data';
-
-export default DS.Model.extend({
-
-  jiraProject: DS.attr('string')
-
-});
