+++ app/serializers/device-preference.js
@@ -1,0 +1,16 @@
+import DRFSerializer from './drf';
+
+export default DRFSerializer.extend({
+  normalizeResponse: function (store, primaryModelClass, payload) {
+    return {
+      data: {
+        id: 1,
+        type: 'device-preference',
+        attributes: {
+          platformVersion: payload.platform_version,
+          deviceType: payload.device_type
+        }
+      }
+    };
+  }
+});
